+++ nebuchadnezzar/nebu/tests/cli/test_publish.py
@@ -1,4 +1,3 @@
-
 import io
 import zipfile
 from cgi import parse_multipart
@@ -154,10 +153,10 @@ class TestPublishCmd:
         mock_successful_ping('/auth-ping', requests_mock)
         mock_successful_ping('/publish-ping', requests_mock)
 
-        id = 'collection_resources'
+        dir_name = 'collection_resources'
         publisher = 'CollegeStax'
         message = 'mEssAgE'
-        monkeypatch.chdir(str(datadir / id))
+        monkeypatch.chdir(str(datadir / dir_name))
         monkeypatch.setenv('XXX_PUBLISHER', publisher)
 
         # Mock the publishing request
@@ -200,16 +199,19 @@ class TestPublishCmd:
         with zipfile.ZipFile(io.BytesIO(form['file'][0])) as zb:
             included_files = set(zb.namelist())
         expected_files = set((
-            'col11405/LEaRN.png',
             'col11405/collection.xml',
+            'col11405/LEaRN.png',
             'col11405/m37154/LEaRN.png',
             'col11405/m37154/index.cnxml',
-            'col11405/m37217/index.cnxml',
-            'col11405/m37386/index.cnxml',
-            'col11405/m40645/index.cnxml',
-            'col11405/m40646/index.cnxml',
-            'col11405/m42303/index.cnxml',
-            'col11405/m42304/index.cnxml'
+            # NOTE: The following files are commented out to create the
+            # expectation that we only publish content that has changed.
+            # See test_publish_only_what_changed for more info.
+            # 'col11405/m37217/index.cnxml',
+            # 'col11405/m37386/index.cnxml',
+            # 'col11405/m40645/index.cnxml',
+            # 'col11405/m40646/index.cnxml',
+            # 'col11405/m42303/index.cnxml',
+            # 'col11405/m42304/index.cnxml',
         ))
         assert included_files == expected_files
 
@@ -535,6 +537,251 @@ class TestPublishCmd:
         assert 'Bad credentials: ' in result.output
         expected_output = (
             'Stop the Press!!! =()\n'
+        )
+        # FIXME Ignoring temporary formatting of output, just check for
+        #       the last line so we know we got to the correct place.
+        # assert result.output == expected_output
+        assert expected_output in result.output
+
+    def test_publish_only_what_changed(self, datadir, monkeypatch,
+                                       requests_mock, invoker):
+        mock_successful_ping('/auth-ping', requests_mock)
+        mock_successful_ping('/publish-ping', requests_mock)
+
+        contents_folder = 'collection_resources'
+        publisher = 'CollegeStax'
+        message = 'mEssAgE'
+        monkeypatch.chdir(str(datadir / contents_folder))
+        monkeypatch.setenv('XXX_PUBLISHER', publisher)
+
+        # Mock the publishing request
+        url = 'https://cnx.org/api/publish-litezip'
+        resp_callback = ResponseCallback(COLLECTION_PUBLISH_PRESS_RESP_DATA)
+        requests_mock.register_uri(
+            'POST',
+            url,
+            status_code=200,
+            text=resp_callback,
+        )
+
+        from nebu.cli.main import cli
+        # Use Current Working Directory (CWD)
+        args = ['publish', 'test-env', '.', '-m', message,
+                '--username', 'someusername', '--password', 'somepassword']
+        result = invoker(cli, args)
+
+        # Check the results
+        if result.exception:
+            raise result.exception
+        assert result.exit_code == 0
+        expected_output = (
+            'Great work!!! =D\n'
+        )
+        # FIXME Ignoring temporary formatting of output, just check for
+        #       the last line so we know we got to the correct place.
+        # assert result.output == expected_output
+        assert expected_output in result.output
+
+        # Check the sent contents
+        request_data = resp_callback.captured_request._request.body
+        # Discover the multipart/form-data boundry
+        boundary = request_data.split(b'\r\n')[0][2:]
+        form = parse_multipart(io.BytesIO(request_data),
+                               {'boundary': boundary})
+        assert form['publisher'][0] == publisher.encode('utf8')
+        assert form['message'][0] == message.encode('utf8')
+        # Check the zipfile for contents
+        with zipfile.ZipFile(io.BytesIO(form['file'][0])) as zb:
+            actual_published = set(zb.namelist())
+        expected_published = set((
+            # Although these commented out files do exist in the directory,
+            # they will not be published because their cached sha1 and
+            # actual/calculated sha1 are equal, meaning there are no changes.
+            # 'col11405/m37217/index.cnxml',
+            # 'col11405/m37386/index.cnxml',
+            # 'col11405/m40645/index.cnxml',
+            # 'col11405/m40646/index.cnxml',
+            # 'col11405/m42303/index.cnxml',
+            # 'col11405/m42304/index.cnxml',
+
+            # This one won't be in the cache (.sha1sum) file
+            # to pretend that it's a new image:
+            'col11405/m37154/LEaRN.png',
+            # This one will have a different sha1 in the cache file
+            # to pretend that it was modified:
+            'col11405/m37154/index.cnxml',
+            # Also this one will have a different sha1 in the cache file
+            # to pretend that it was modified:
+            'col11405/LEaRN.png',
+            # Since modules were modified, also assume the collection
+            # was modified:
+            'col11405/collection.xml',
+        ))
+
+        assert actual_published == expected_published
+
+    def test_publish_cmd_with_no_cache(self, datadir, monkeypatch,
+                                       requests_mock, invoker):
+        """Test what happens when there is no .sha1sum file in any dir.
+        """
+        mock_successful_ping('/auth-ping', requests_mock)
+        mock_successful_ping('/publish-ping', requests_mock)
+
+        contents_folder = 'collection'
+        publisher = 'CollegeStax'
+        message = 'mEssAgE'
+        monkeypatch.chdir(str(datadir / contents_folder))
+        monkeypatch.setenv('XXX_PUBLISHER', publisher)
+
+        # Mock the publishing request
+        url = 'https://cnx.org/api/publish-litezip'
+        resp_callback = ResponseCallback(COLLECTION_PUBLISH_PRESS_RESP_DATA)
+        requests_mock.register_uri(
+            'POST',
+            url,
+            status_code=200,
+            text=resp_callback,
+        )
+
+        from nebu.cli.main import cli
+        # Use Current Working Directory (CWD)
+        args = ['publish', 'test-env', '.', '-m', message,
+                '--username', 'someusername', '--password', 'somepassword']
+        result = invoker(cli, args)
+
+        # Check the results
+        if result.exception:
+            raise result.exception
+        assert result.exit_code == 0
+        expected_output = (
+            'Great work!!! =D\n'
+        )
+        # FIXME Ignoring temporary formatting of output, just check for
+        #       the last line so we know we got to the correct place.
+        # assert result.output == expected_output
+        assert expected_output in result.output
+
+        # Check the sent contents
+        request_data = resp_callback.captured_request._request.body
+        # Discover the multipart/form-data boundry
+        boundary = request_data.split(b'\r\n')[0][2:]
+        form = parse_multipart(io.BytesIO(request_data),
+                               {'boundary': boundary})
+        assert form['publisher'][0] == publisher.encode('utf8')
+        assert form['message'][0] == message.encode('utf8')
+        # Check the zipfile for contents
+        with zipfile.ZipFile(io.BytesIO(form['file'][0])) as zb:
+            actual_published = set(zb.namelist())
+
+        # There is no .sha1sum file in any of the directories within
+        # data/collection folder, so it should treat all files as new and
+        # publish them all.
+        expected_published = set((
+            'col11405/collection.xml',
+            'col11405/m37151/index.cnxml',
+            'col11405/m37152/index.cnxml',
+            'col11405/m37154/index.cnxml',
+            'col11405/m37217/index.cnxml',
+            'col11405/m37386/index.cnxml',
+            'col11405/m40643/index.cnxml',
+            'col11405/m40645/index.cnxml',
+            'col11405/m40646/index.cnxml',
+            'col11405/m42302/index.cnxml',
+            'col11405/m42303/index.cnxml',
+            'col11405/m42304/index.cnxml'
+        ))
+        assert actual_published == expected_published
+
+    def test_publish_only_what_changed_with_tree(self, datadir, monkeypatch,
+                                                 requests_mock, invoker):
+        mock_successful_ping('/auth-ping', requests_mock)
+        mock_successful_ping('/publish-ping', requests_mock)
+
+        contents_folder = 'book_tree_cached'
+        publisher = 'CollegeStax'
+        message = 'mEssAgE'
+        monkeypatch.chdir(str(datadir / contents_folder))
+        monkeypatch.setenv('XXX_PUBLISHER', publisher)
+
+        # Mock the publishing request
+        url = 'https://cnx.org/api/publish-litezip'
+        resp_callback = ResponseCallback(COLLECTION_PUBLISH_PRESS_RESP_DATA)
+        requests_mock.register_uri(
+            'POST',
+            url,
+            status_code=200,
+            text=resp_callback,
+        )
+
+        from nebu.cli.main import cli
+        # Use Current Working Directory (CWD)
+        args = ['publish', 'test-env', '.', '-m', message,
+                '--username', 'someusername', '--password', 'somepassword']
+        result = invoker(cli, args)
+
+        # Check the results
+        if result.exception:
+            raise result.exception
+        assert result.exit_code == 0
+        expected_output = (
+            'Great work!!! =D\n'
+        )
+        # FIXME Ignoring temporary formatting of output, just check for
+        #       the last line so we know we got to the correct place.
+        # assert result.output == expected_output
+        assert expected_output in result.output
+
+        # Check the sent contents
+        request_data = resp_callback.captured_request._request.body
+        # Discover the multipart/form-data boundry
+        boundary = request_data.split(b'\r\n')[0][2:]
+        form = parse_multipart(io.BytesIO(request_data),
+                               {'boundary': boundary})
+        assert form['publisher'][0] == publisher.encode('utf8')
+        assert form['message'][0] == message.encode('utf8')
+        # Check the zipfile for contents
+        with zipfile.ZipFile(io.BytesIO(form['file'][0])) as zb:
+            actual_published = set(zb.namelist())
+        expected_published = set((
+            'col11405/collection.xml',
+            'col11405/m42302/index.cnxml',
+        ))
+
+        assert actual_published == expected_published
+
+    def test_nothing_changed(self, datadir, monkeypatch,
+                             requests_mock, invoker):
+        mock_successful_ping('/auth-ping', requests_mock)
+        mock_successful_ping('/publish-ping', requests_mock)
+
+        contents_folder = 'collection_no_changes'
+        publisher = 'CollegeStax'
+        message = 'mEssAgE'
+        monkeypatch.chdir(str(datadir / contents_folder))
+        monkeypatch.setenv('XXX_PUBLISHER', publisher)
+
+        # Mock the publishing request
+        url = 'https://cnx.org/api/publish-litezip'
+        resp_callback = ResponseCallback(COLLECTION_PUBLISH_PRESS_RESP_DATA)
+        requests_mock.register_uri(
+            'POST',
+            url,
+            status_code=200,
+            text=resp_callback,
+        )
+
+        from nebu.cli.main import cli
+        # Use Current Working Directory (CWD)
+        args = ['publish', 'test-env', '.', '-m', message,
+                '--username', 'someusername', '--password', 'somepassword']
+        result = invoker(cli, args)
+
+        # Check the results
+        if result.exception and not isinstance(result.exception, SystemExit):
+            raise result.exception
+        assert result.exit_code == 1
+        expected_output = (
+            'Nothing changed, nothing to publish.\n'
         )
         # FIXME Ignoring temporary formatting of output, just check for
         #       the last line so we know we got to the correct place.